refactor(og): move OgCardGenerator under global/og/service
Aligns with OgController location (global/og/controller) — og image
concerns now live together rather than nested under workspace.
